requirements.txt
flask
flask-cors
torch
transformers
sentence-transformers
faiss-cpu
pypdf
langchain
nltk
nltk